do some package refactoring
