+++ app/src/main/java/io/mstream/mstream/BaseActivity.java
@@ -22,7 +22,10 @@ import android.widget.Toast;
 import java.util.LinkedList;
 import java.util.List;
 
+import io.mstream.mstream.filebrowser.FileBrowserFragment;
 import io.mstream.mstream.filebrowser.FileItem;
+import io.mstream.mstream.player.JukeboxService;
+import io.mstream.mstream.playlist.PlaylistFragment;
 import io.mstream.mstream.serverlist.ServerItem;
 import io.mstream.mstream.serverlist.ServerListAdapter;
 import io.mstream.mstream.serverlist.ServerStore;
