+++ app/src/main/res/layout/fragment_file_browser.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="io.mstream.mstream.FileBrowserFragment">
+    tools:context="io.mstream.mstream.filebrowser.FileBrowserFragment">
 
     <!-- TODO: Update blank fragment layout -->
