+++ app/src/main/res/layout/fragment_playlist.xml
@@ -6,4 +6,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center_horizontal|top"
-    tools:context="io.mstream.mstream.PlaylistFragment"/>
+    tools:context="io.mstream.mstream.playlist.PlaylistFragment"/>
